BUG#37697570 BUG#37697649 Defining RAM usage limits in jit-executor
By defining minimum and maximum allowed RAM limits the following
issues get resolved:

- BUG#37697649 infinite loop at GET mrs script
- BUG#37697570 too large maximumRamUser leads to HTTP 501

The ram to be used by the jit-executor is defined as follows:

- Default RAM: 25% of system memory
- Maximum Allowed: 80% of system memory
- Minimum Allowed: 10 MB

If configured RAM is above the max allowed, the jit-executor will
run with defaults, which means it will use as much as needed with
a limit of 80% of the system memory.

If configured RAM is below the min allowed, the jit-executor will
use the default value of 25% of the system memory.

+++ router/src/jit_executor/src/jit_executor_component.cc
@@ -149,10 +149,13 @@ void JitExecutorComponent::update_active_contexts(
   // Total pool is verified here as it could be 0, meaning, no active context
   // handlers will be left after this update
   if (total_pool != 0 && m_global_config.maximum_ram_size.has_value()) {
-    uint64_t mem_per_pool_item = *m_global_config.maximum_ram_size / total_pool;
+    // Use double to get the most accurate value per pool item
+    double mem_per_pool_item =
+        *m_global_config.maximum_ram_size / static_cast<double>(total_pool);
 
     for (const auto &it : candidate_context_handlers) {
-      it.second->set_max_heap_size(mem_per_pool_item * it.second->pool_size());
+      it.second->set_max_heap_size(
+          static_cast<uint64_t>(mem_per_pool_item * it.second->pool_size()));
     }
   }
